fix: Always update selection on a re-render
This will allow it to be set on an auto-recover of the parent form

Previously, the selection would be set to `[]` on mount, and then
an auto-recover would change the form field (updating the field value),
but not update the live_select selection, because it was already set.

This couldn't be avoided by setting the `value` for the live_select
component because that triggers a client event, which seems to invoke
an infinite loop between the `phx-blur` and `phx-change` events
defined for the component, when invoked after an auto-recover.