Bump live_view to 1.0
LiveView 1.0 adds support for interpolation using curly braces (`{...}`) instead
of angle brackets (`<%= ... %>`)[0]. This caused a compilation error in
`ShowcaseLive`, because we return literal curly braces in the HTML. The fix is to
either use `&lbrace;` or `<%= "{" %>`[1]. I think the former is nicer, so that's
what I've done here. If you disagree, I'm happy to change it.

I've also changed some instances of `live_flash/2` to `Phoenix.Flash.get/2`, as
`live_flash/2` is now deprecated[2]

Fixes: #88

0: https://hexdocs.pm/phoenix_live_view/Phoenix.Component.html#sigil_H/2-interpolation

1: https://github.com/phoenixframework/phoenix_live_view/pull/3554

2: https://hexdocs.pm/phoenix_live_view/changelog.html#deprecations-1