Clicking disabled options won't close the dropdown
This modifies the behavior in Tags and Single mode for
clicking a disabled object. Previously clicking one in either mode
would close the dropdown, but leave the user focused in the input.

This potentially gave an appearance to the User that their input
was accepted. So instead of closing the option dropdown, it's left open
for the user to interact with like a regular select would with disabled
options.