Unify blur experience in single and multi modes
Currently blur behaves differently depending on mode, for single it hides user input if nothing was selected but it doesn't do the same for tags. 

This commit makes single with selection the only special case where we show something on blur (the selected label) .. otherwise just hide user input in all other cases.